+++ Assets/Resources/Story1TextScript.csv
@@ -1,0 +1,31 @@
+﻿Talk,Name,Text,Delay,
+1,,이건 내가 어렸을 적에 있던 일이다.,1,
+2,,많은 집들이 그러하듯이 우리 집도 맞벌이였다.,1,
+3,,"이 일이 있던 시기는 내가 초등학생 저학년쯤 되었을 때 일인데, 아직까지도 난 이 일을 잊을 수 없다.",1,
+4,,"그때로부터 많은 시간들이 지났지만, 그때 느꼈던 감정들, 그 공포들을 현재 성인이 되어서도 내 몸을 감싼다.",1,
+5,,"집에 어른이라곤 아무도 없었고, 단지 나 혼자서…그 많은 공포들을 감당했어야 됬다.",1,
+6,,이 일의 시작은 부모님이 외롭게 혼자 있는 나를 위해 인형 3개를 사오가지고 오시면서 시작되었다.,1,
+7,,,1,
+8,,이 3개의 인형이 바로 날 공포로 몰아넣은 주범들이다. ,1,
+9,,보기에는 그냥 평범한 인형이라고 생각할 수 있다만… ,1,
+10,,저주 받은 인형들이다.,1,
+11,,"무슨일들이 있었냐 하면….. 아무도 모르게 물건들의 위치가 바뀌어 있다던가, 인형의 위치가 달라진다던가,",1,
+12,,아니면…일어났을 때 저 멀리서 날 바라보고 있다던가….사실 이정도면 다행이다,1,
+13,,어느 날 부턴 알 수 없는 상처가 생기기 시작했고 나 뿐만 아니라 부모님도 같은 일들을 겪었다.,1,
+14,,결국 유명한 심령술사를 불렀고 그는 이와 같이 말했다,1,
+15,심령술사,음….좋지 않은 기운이 집 전체를 감싸고 있군요…,1,
+16,심령술사,이 집은 저주받았네요… 어느정도 조치를 하겠지만…그냥 이곳을 벗어나는게 가장 좋을 겁니다. 아니면 태우든가...,1,
+17,,그렇게 하곤 심령술사는 이상한 술식을 하고 나왔다. 하지만…,1,
+18,,몇일 뒤 그 심령술사가 죽었다는 뉴스를 듣고 우리 가족 모두 놀랐다.,1,
+19,,아… 그만큼 현재 저주가 엄청나구나',1,
+20,,코스믹 호러라고 말하던가…인간이 어떻게 할 수 없는 무언가를 만났을 때 느끼는 무력감과 비슷한 공포…,1,
+21,,하지만 우리의 머리속을 통과한건 심령술사가 말했던 무언가였다,1,
+22,심령술사,아니면…태우든가,1,
+23,,우리 가족 모두가 알고 있었다…그…..그…..저주가 시작된 이유를..,1,
+24,,어렸던 나는 생각했다,1,
+25,,"""인형을 불태우자""",1,
+26,,허나 인형의 모습을 찾을 수 없었다. 마치 자신을 찾아보라는 것처럼..우리의 생각을 이미 알고 있는 것처럼,1,
+27,,하지만 ….그렇다고 해서 포기할 수 없었다.. 어떻게든 찾아야 해!,1,
+28,,일단 이 방에서 인형을 찾아보자',1,
+29,,문은 잠겨 있는 것 같다',1,
+30,,,1,
