+++ Assets/Resources/Story1TextScript.csv.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 030ac18b38f8681439d6f89a47272864
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
