+++ Assets/Scenes/Story4Scene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7953cc1f10259ee48b9a7749dc43f68a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
