--- Assets/Scripts/UI/ButtonTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6a11ec48441ccb64a8696c8d8df9f125
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
