+++ Assets/Scripts/UI/Buttons/ButtonHitArea.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class ButtonHitArea : MonoBehaviour
+{
+    void Start()
+    {
+        GetComponent<Image>().alphaHitTestMinimumThreshold = 0.1f;
+    }
+}
