+++ Assets/Scripts/UI/Buttons/ButtonHitArea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f3d3ed7ef987894b95686157833ae9b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
