+++ Assets/Scripts/UI/Buttons/Story1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 400c7e3d22ead384b8d7129252130a65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
