--- Assets/Scripts/UI/FadeUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c6a5e3d295703734fbac4f1d46c11ee5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
