--- Assets/Scripts/UI/Panel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aa787f3c83bf0cb47ac9b43f596b41a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
