+++ Assets/Scripts/UI/Singletons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 130f61f1be6983a4b9f4577d445f72f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
