+++ Assets/Scripts/UI/Texts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 616e2f7be0083224d9b91e7725914fa7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
