+++ Assets/Scripts/UI/Texts/NPCTextUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 087585273cdf082408da857229cf93bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
