--- Assets/Scripts/UI/TitleUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3b3ffbdf49c1094b971f91f66733226
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
