+++ Assets/Scripts/UI/Titles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d139c65700355f4fbf62f4cc8cfd8d9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
