+++ Assets/TextMesh Pro/Fonts/Pretendard-Light.ttf.meta
@@ -1,0 +1,27 @@
+fileFormatVersion: 2
+guid: fedcae9adef04d1438adf38ddfcb3f47
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - Pretendard
+  fallbackFontReferences:
+  - {fileID: 12800000, guid: 4c6378a97493fce438c1044160ef08d9, type: 3}
+  - {fileID: 12800000, guid: 34fa108cab80362429931364c70d8f49, type: 3}
+  - {fileID: 12800000, guid: 8d5df0795a8f5694bafee81e72d71f80, type: 3}
+  - {fileID: 12800000, guid: fcc0ad41daba2714e97b5a2dce10704c, type: 3}
+  - {fileID: 12800000, guid: dabe52f33a56a4643a791d704499cb93, type: 3}
+  - {fileID: 12800000, guid: 9d1c3292ac1ce2c42b9fe9665be36cff, type: 3}
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
