+++ Assets/TextMesh Pro/Fonts/Pretendard-Medium SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f113f2306adef84bb9cf4aa8ee41ac2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
