+++ ProjectSettings/EditorBuildSettings.asset
@@ -12,7 +12,7 @@ EditorBuildSettings:
     path: Assets/Scenes/TitleScene.unity
     guid: 3c161f8f40202db43b0d1c02b5e8fbe4
   - enabled: 1
-    path: Assets/Scenes/StoryBridgeScene.unity
+    path: Assets/Scenes/Story4Scene.unity
     guid: 7953cc1f10259ee48b9a7749dc43f68a
   - enabled: 1
     path: Assets/Scenes/Story1Scene.unity
